\name{KeywordsRegistor.run}
\alias{KeywordsRegistor.run}
\title{Keywords Registor}
\usage{
KeywordsRegistor.run()
}
\description{
Run the shiny app "Keywords Registor". Register clusterings completed on an enriched fcs file without using the "Clustering Tool"" to make them compatible with the "Analysis Tool".
}
\examples{
%YOU MIGHT HAVE TO LOAD ANOTHER R SESSION BEFORE USING THE TOOL
library(KeywordsRegistor)
KeywordsRegistor.run()
}
